refactor(bridge): extract bridge monitoring RPC calls to a separate module
