+++ backend/crates/bridge/src/status.rs
@@ -1,19 +1,14 @@
 use anyhow::Result;
 use axum::Json;
 use bitcoin::{secp256k1::PublicKey, Txid};
-
-use jsonrpsee::http_client::HttpClient;
-use std::collections::BTreeMap;
 use std::sync::Arc;
-use strata_bridge_rpc::traits::{StrataBridgeControlApiClient, StrataBridgeMonitoringApiClient};
-use strata_bridge_rpc::types::{
-    RpcDepositStatus, RpcOperatorStatus, RpcReimbursementStatus, RpcWithdrawalStatus,
-};
+use strata_bridge_rpc::types::{RpcDepositStatus, RpcReimbursementStatus, RpcWithdrawalStatus};
 
 use strata_primitives::buf::Buf32;
 use strata_tasks::ShutdownGuard;
 
 use super::{
+    bridge_rpc::{self, RpcClientManager},
     context::BridgeMonitoringContext,
     types::{
         BridgeStatus, DepositInfo, DepositStatus, OperatorStatus, ReimbursementInfo,
@@ -22,140 +17,9 @@ use super::{
 };
 
 use tokio::time::{interval, Duration};
-use tracing::{debug, error, info, warn};
+use tracing::{error, info, warn};
 
 use status_config::BridgeMonitoringConfig;
-use status_utils::{create_rpc_client, execute_with_retries};
-
-/// RPC client manager with connection pooling and retry logic
-///
-/// This manager maintains a pool of reusable HTTP clients for each bridge operator,
-/// preventing connection exhaustion by reusing clients across requests. It implements
-/// automatic retry logic with exponential backoff to handle transient network failures.
-///
-/// # Design
-///
-/// - **Connection Pooling**: Creates one HTTP client per operator and reuses it for all requests
-/// - **Retry Logic**: Implements exponential backoff (3 retries over 10 seconds with 1.5x multiplier)
-/// - **Failover**: Tries all available operators in deterministic order (sorted by public key)
-/// - **Graceful Degradation**: Returns `None` if all operators fail after retries
-///
-/// # Example Flow
-///
-/// For each RPC request:
-///
-/// 1. Try operator 1 with up to 3 retries (exponential backoff between retries)
-/// 2. If operator 1 fails after retries, try operator 2 with up to 3 retries
-/// 3. Continue until an operator succeeds or all fail
-/// 4. Return the first successful result or [`None`] if all fail
-struct RpcClientManager {
-    /// HTTP clients for each operator, keyed by operator public key ([`String`])
-    /// [`BTreeMap`] ensures deterministic ordering (sorted by key)
-    clients: BTreeMap<String, HttpClient>,
-}
-
-impl RpcClientManager {
-    /// Create a new RPC client manager for the configured bridge operators
-    ///
-    /// This initializes one HTTP client per operator with:
-    ///
-    /// - 30-second request timeout
-    /// - 10MB max request size
-    /// - Connection pooling enabled
-    ///
-    /// # Arguments
-    ///
-    /// * `config` - Bridge monitoring configuration containing operator RPC URLs
-    fn new(config: &BridgeMonitoringConfig) -> Self {
-        let mut clients = BTreeMap::new();
-        for operator in config.operators() {
-            clients.insert(
-                operator.public_key().to_string(),
-                create_rpc_client(operator.rpc_url()),
-            );
-        }
-
-        Self { clients }
-    }
-
-    /// Execute an async operation across all available clients with retry logic
-    ///
-    /// This method tries the given operation on each operator sequentially (in sorted order
-    /// by public key). For each operator, it retries up to 3 times with exponential
-    /// backoff between attempts using [`execute_with_retries`]. The first successful result
-    /// is returned immediately.
-    ///
-    /// # Type Parameters
-    ///
-    /// * `T` - The expected return type of the operation
-    /// * `F` - A function that takes an [`HttpClient`] and returns a future
-    /// * `Fut` - The future returned by `F`
-    ///
-    /// # Arguments
-    ///
-    /// * `operation` - A closure that performs an RPC call using the provided client
-    ///
-    /// # Returns
-    ///
-    /// * [`Some(T)`](Some) - If any operator succeeds (possibly after retries)
-    /// * [`None`] - If all operators fail after exhausting their retries
-    ///
-    /// # Retry Behavior
-    ///
-    /// Uses [`execute_with_retries`] for each operator with exponential backoff:
-    ///
-    /// - **Attempt 0**: Immediate (no delay)
-    /// - **Attempt 1**: After ~2s delay
-    /// - **Attempt 2**: After ~3s delay  
-    /// - **Attempt 3**: After ~5s delay
-    /// - Total: ~10 seconds per operator
-    ///
-    /// # Example
-    ///
-    /// ```ignore
-    /// let result = rpc_manager
-    ///     .query_clients_with_retry(|client| async move {
-    ///         client.get_deposit_requests().await.map_err(|e| e.into())
-    ///     })
-    ///     .await;
-    /// ```
-    async fn query_clients_with_retry<T, F, Fut>(&self, operation: F) -> Option<T>
-    where
-        F: Fn(HttpClient) -> Fut,
-        Fut: std::future::Future<Output = Result<T, Box<dyn std::error::Error + Send + Sync>>>,
-    {
-        // BTreeMap maintains sorted order automatically
-        for (key, client) in self.clients.iter() {
-            let client_clone = client.clone();
-            let operation_name = format!("RPC request to operator public key {key}");
-
-            match execute_with_retries(
-                || {
-                    let client = client_clone.clone();
-                    operation(client)
-                },
-                &operation_name,
-            )
-            .await
-            {
-                Ok(result) => {
-                    debug!(operator_pk = %key, "rpc request succeeded");
-                    return Some(result);
-                }
-                Err(e) => {
-                    warn!(
-                        operator_pk = %key,
-                        error = %e,
-                        "rpc request failed after retries"
-                    );
-                    // Continue to next operator
-                }
-            }
-        }
-
-        None
-    }
-}
 
 /// Get transaction confirmations from esplora
 async fn get_tx_confirmations(esplora_url: &str, txid: Txid, chain_tip_height: u64) -> Option<u64> {
@@ -300,7 +164,7 @@ pub async fn bridge_monitoring_task(
             let operator_id = format!("Alpen Labs #{}", index + 1);
             let pk_bytes = hex::decode(operator.public_key()).expect("decode to succeed");
             let operator_pk = PublicKey::from_slice(&pk_bytes).expect("conversion to succeed");
-            let status = get_operator_status(operator.rpc_url()).await;
+            let status = bridge_rpc::get_operator_status(operator.rpc_url()).await;
             operator_statuses.push(OperatorStatus::new(operator_id, operator_pk, status));
         }
 
@@ -451,19 +315,6 @@ pub async fn bridge_monitoring_task(
     Ok(())
 }
 
-/// Fetch operator status
-async fn get_operator_status(rpc_url: &str) -> RpcOperatorStatus {
-    let rpc_client = create_rpc_client(rpc_url);
-
-    // Directly use `get_uptime`
-    if rpc_client.get_uptime().await.is_ok() {
-        RpcOperatorStatus::Online
-    } else {
-        warn!("failed to fetch bridge operator uptime");
-        RpcOperatorStatus::Offline
-    }
-}
-
 /// Fetch bitcoin chain tip height
 async fn get_bitcoin_chain_tip_height(
     esplora_url: &str,
@@ -474,32 +325,6 @@ async fn get_bitcoin_chain_tip_height(
     let text = resp.text().await?;
     let height = text.trim().parse::<u64>()?;
     Ok(height)
-}
-
-/// Fetch all pending deposit request transaction IDs from bridge operators
-///
-/// Uses the RPC client manager with retry logic to query all operators for their
-/// list of deposit requests. Returns the first non-empty result, or an empty vector
-/// if all operators have no deposits or all fail.
-///
-/// # Arguments
-///
-/// * `rpc_manager` - RPC client manager with retry/failover logic
-///
-/// # Returns
-///
-/// Vector of deposit request transaction IDs. Empty if no deposits found or all operators failed.
-async fn get_deposit_requests(rpc_manager: &RpcClientManager) -> Vec<Txid> {
-    let result = rpc_manager
-        .query_clients_with_retry(|client| async move {
-            client.get_deposit_requests().await.map_err(|e| e.into())
-        })
-        .await;
-
-    result.unwrap_or_else(|| {
-        warn!("no deposit requests found from any operator");
-        Vec::new()
-    })
 }
 
 /// Fetch detailed information for all deposit requests
@@ -529,7 +354,7 @@ async fn get_deposits(
     chain_tip_height: u64,
     active_deposit_txids: &[Txid],
 ) -> Vec<(DepositInfo, u64)> {
-    let new_deposit_requests = get_deposit_requests(rpc_manager).await;
+    let new_deposit_requests = bridge_rpc::get_deposit_requests(rpc_manager).await;
     let new_count = new_deposit_requests.len();
     let mut deposit_requests = new_deposit_requests;
 
@@ -550,14 +375,7 @@ async fn get_deposits(
     let mut deposit_infos: Vec<(DepositInfo, u64)> = Vec::new();
     for deposit_request_txid in deposit_requests.iter() {
         let txid = *deposit_request_txid;
-        let rpc_info = rpc_manager
-            .query_clients_with_retry(|client| async move {
-                client
-                    .get_deposit_request_info(txid)
-                    .await
-                    .map_err(|e| e.into())
-            })
-            .await;
+        let rpc_info = bridge_rpc::get_deposit_request_info(rpc_manager, txid).await;
 
         let Some(dep_info) = rpc_info else {
             error!(%deposit_request_txid, "failed to fetch deposit info after retries");
@@ -594,32 +412,6 @@ async fn get_deposits(
     deposit_infos
 }
 
-/// Fetch all pending withdrawal request IDs from bridge operators
-///
-/// Uses the RPC client manager with retry logic to query all operators for their
-/// list of withdrawal requests. Returns the first non-empty result, or an empty vector
-/// if all operators have no withdrawals or all fail.
-///
-/// # Arguments
-///
-/// * `rpc_manager` - RPC client manager with retry/failover logic
-///
-/// # Returns
-///
-/// Vector of withdrawal request IDs. Empty if no withdrawals found or all operators failed.
-async fn get_withdrawal_requests(rpc_manager: &RpcClientManager) -> Vec<Buf32> {
-    let result = rpc_manager
-        .query_clients_with_retry(|client| async move {
-            client.get_withdrawals().await.map_err(|e| e.into())
-        })
-        .await;
-
-    result.unwrap_or_else(|| {
-        warn!("no withdrawal requests found from any operator");
-        Vec::new()
-    })
-}
-
 /// Fetch detailed information for all withdrawal requests and fulfillments
 ///
 /// Queries bridge operators for withdrawal details, including both new withdrawal requests
@@ -646,7 +438,7 @@ async fn get_withdrawals(
     chain_tip_height: u64,
     active_withdrawal_request_ids: &[Buf32],
 ) -> Vec<(WithdrawalInfo, u64)> {
-    let new_withdrawal_requests = get_withdrawal_requests(rpc_manager).await;
+    let new_withdrawal_requests = bridge_rpc::get_withdrawal_requests(rpc_manager).await;
     let new_count = new_withdrawal_requests.len();
     let mut withdrawal_requests = new_withdrawal_requests;
 
@@ -667,15 +459,7 @@ async fn get_withdrawals(
     let mut withdrawal_infos: Vec<(WithdrawalInfo, u64)> = Vec::new();
     for withdrawal_request_txid in withdrawal_requests.iter() {
         let request_id = *withdrawal_request_txid;
-        let rpc_info = rpc_manager
-            .query_clients_with_retry(|client| async move {
-                match client.get_withdrawal_info(request_id).await {
-                    Ok(Some(info)) => Ok(info),
-                    Ok(None) => Err("No withdrawal info found".into()),
-                    Err(e) => Err(e.into()),
-                }
-            })
-            .await;
+        let rpc_info = bridge_rpc::get_withdrawal_info(rpc_manager, request_id).await;
 
         let Some(wd_info) = rpc_info else {
             error!(%withdrawal_request_txid, "failed to fetch withdrawal info after retries");
@@ -707,32 +491,6 @@ async fn get_withdrawals(
     withdrawal_infos
 }
 
-/// Fetch all pending claim transaction IDs from bridge operators
-///
-/// Uses the RPC client manager with retry logic to query all operators for their
-/// list of claim transactions. Returns the first non-empty result, or an empty vector
-/// if all operators have no claims or all fail.
-///
-/// # Arguments
-///
-/// * `rpc_manager` - RPC client manager with retry/failover logic
-///
-/// # Returns
-///
-/// Vector of claim transaction IDs. Empty if no claims found or all operators failed.
-async fn get_claims(rpc_manager: &RpcClientManager) -> Vec<Txid> {
-    let result = rpc_manager
-        .query_clients_with_retry(|client| async move {
-            client.get_claims().await.map_err(|e| e.into())
-        })
-        .await;
-
-    result.unwrap_or_else(|| {
-        warn!("no claims found from any operator");
-        Vec::new()
-    })
-}
-
 /// Fetch detailed information for all claim and reimbursement transactions
 ///
 /// Queries bridge operators for claim/reimbursement details, including both new claims
@@ -760,7 +518,7 @@ async fn get_reimbursements(
     chain_tip_height: u64,
     active_reimbursement_txids: &[Txid],
 ) -> Vec<(ReimbursementInfo, u64)> {
-    let new_claims = get_claims(rpc_manager).await;
+    let new_claims = bridge_rpc::get_claims(rpc_manager).await;
     let new_count = new_claims.len();
     let mut claims = new_claims;
 
@@ -781,15 +539,7 @@ async fn get_reimbursements(
     let mut reimbursement_infos = Vec::new();
     for claim_txid in claims.iter() {
         let txid = *claim_txid;
-        let rpc_info = rpc_manager
-            .query_clients_with_retry(|client| async move {
-                match client.get_claim_info(txid).await {
-                    Ok(Some(info)) => Ok(info),
-                    Ok(None) => Err("No claim info found".into()),
-                    Err(e) => Err(e.into()),
-                }
-            })
-            .await;
+        let rpc_info = bridge_rpc::get_claim_info(rpc_manager, txid).await;
 
         let Some(claim_info) = rpc_info else {
             error!(%claim_txid, "failed to fetch claim info after retries");
